Some Code for Driving, Impractical to Continue
I found that there are 3 different files that have to be edited , the main.h, the auto.c and the init.c to be able to use the red encoder. The IME is much easier, and therefore complete for drive.  I determined that it would be impractical and not worth my time to continue development, but am posting my current work anyway. I'm sorry to all who would like to use this on PROS, but feel free to fork this and continue if you wish!